Refactor all crypto functionality to 'Crypto' struct
